+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+* SymuEngine - delete EnvironmentEntity => OrganizationModels by [@lmorisse]
 
 ## 0.2.0 [☰](https://github.com/lmorisse/symu/compare/v0.2.0..v0.1.0)
 * Add new project example: [SymuLearnAndForget](https://github.com/lmorisse/Symu/tree/master/Symu%20examples/SymuLearnAndForget) by [@lmorisse]